Updated JovoModelHelper
The JovoModelHelper has more methods to access more parts of the JovoLanguageModel.
Also a JovoModelBuilder was implemented, which uses all methods of the JovoModelHelper. The implementations are loaded when a new builder is constructed. This makes it easier to maintain, because only the JovoModelHelper has to be adjusted and maybe definitions added to the JovoModelBuilder.